Begin style conformance with new Awesome List guides.
The Awesome List style guide wants listed awesome lists to conform to a
particular style. Their style guide includes numerous items, some of
which are addressed by this commit. The items addressed herein include:

* Name of list as the top-level heading.
* Awesome badge placed to the right of the title, not underneath.
* Description of the list contents.
* All line items should end in appropriate punctuation (like a period).

This is a first pass and should not be considered an exhaustive
accounting of the style guide's recommendations.